(#21427) Make report format selectable
Without making agents able to select the format to serialize reports, a
newer agent would be unable to submit reports to an older master.
Although we ask users to upgrade masters first, that is not often done
in practice. In order to allow those kinds of setups to work, this makes
the report format selectable.

Puppet 3.2.2 and later support reports as pson, so only older, insecure
masters need to have the agent's change this new setting.