(#21427) Handle loss of precision during YAML serialization
Previously, when round-tripping a report using YAML, the
Puppet::Resource::Status#time object could lose precision (depending on
the ruby version and platform), so it did not match the expected value. For
example, the expected value was:

    "time"=>"2013-09-04T17:01:38.510604415+00:00"

But the round-tripped value was:

    "time"=>"2013-09-04T17:01:38.510604000+00:00"

This commit changes the test to explicitly compare status fields, and
handles the `time` field specially to avoid loss of precision during
YAML serialization.