(#22383) Allow all boolean values for atboot
The previous restrictions placed on the atboot property for mount caused
some previously valid manifests to begin to fail. The problem stems from
the fact that atboot is only used on solaris, but can always be
specified by users. It used to be able to be nearly anything, but only
yes and no meant anything on solaris. This fixes the problem by
expanding the allowed values to include all of the "boolean" values as
expressed in the Puppet::[Parameter|Property]::Boolean type and using a
munge method to convert the boolean to the correct yes or no value.