(maint) Handle repository configuration for redhat and debian
Previous harness assumed el-6-x86_64 for both master and agent.  This
change gets all of puppetlabs repository information for the constructed
puppet SHA for both redhat and debian and combines it in a single tarball.
The install setup test then handles extracting and installing the specific
configurations for the particular redhat/debian version based on the
host configuration.

The repos.tar can be pre-loaded so that it is only created once and then
bundled in the acceptance-artifacts.tar.gz for matrix jobs.