(#22442) Fix problem with missing issue ILLEGAL_NAME (future parser)
When a qualified name did not conform to the naming rule for
qualified names a reference was made to an undefined issue.

There was also a problem with the name validation rule as it was not
anchored at the beginning of the matched text. It thus allowed
variable names like `$Aaa` (which should also have been flagged as 
an error). Other naming validation patterns suffered from the same
problem.