(#22356) validate json schemas against the json meta-schema
The recently added json schema were not themselves being
validated.  And unfortunately, the json-schema gem doesn't
do any validation of the input schema, but rather is prone to
raising exceptions if the schema is invalid.

So add a spec test per-schema which validates the schema against
the meta-schema.

Also, fix up resource_type.json so that it is valid.