(#22324) Generalize distmoduledir setting for windows module install
The pe test changes were using a Beaker host setting
agent['distmoduledir'] which was not set for hosts of type Windows.
Furthermore, the actual value is dependent on confdir which varies
by verison according to COMMON_APPDATA, and so is not practicle to
in the base Beaker Windows host settings.  So I'm setting it based
the returned confdir setting from the puppet agent on the host.