(#22535) add backward compatibility for query parameters
Background: in 3.3 yaml was deprecated across the network. One
occurence of yaml across the network was the serialization of
an array into query parameters, notably as occurs for the ignore
list in a file resource (which crosses the wire as a query parameter
to the file_metadata endpoint).

In 3.3.0 this yaml serialization was competely removed.  This
commit adds a puppet option to restore yaml serialization of query
parameters, which allows a newer (3.3.1 <= version < 4) agent to
talk to an older master (3.0.0 <= version 3.2.x).