(#22324) Ensure windows certs are properly removed and regenerated
There are two issues in this test on windows. The first involves spaces
in the path to the ssldir silently preventing us from removing the ssl
dir unless we quote it.  This causes the window's agents certs to fall
out of sync with the master.

The second issue has to do with the confine and Beaker modifying the
testcases host array such that the teardown reset of the ssl was also
confined not to run on windows and again we leave the test with certs
out of sync on windows and failing in subsequent tests that rely on
master/agent communication.

The fix here is to quote the ssldir path, and guard the confine so that
we recover the full array of hosts for the teardown.