(#22535) Add puppet version to status indirector
Without this change, there was no way to query a puppet node for
what version of puppet code it was running.  This change adds a version
field to the status hash.

This allows either a local or rest find on the status indirection
to return the version of puppet code.  This is especially useful
for a client communicating via the REST API.