(#22535) Revert to yaml when a new agent talks to old master
With the 3.3.0 release yaml was deprecated.  However, a 3.3 agent
talking to an older master wasn't fully operational.  This change
causes a 3.3.1+ agent to drop back to yaml when talking with an
older master.  It also issues cautionary messages since the Right
Thing to do in this situation is to upgrade the puppet master.