(Maint) Don't keep Tempfile open while specs are running
Previously, the specs were creating a Tempfile instance in a before
block, writing to the file, but keeping the handle open, until the
after each block ran. This was done to ensure the Tempfile wasn't
deleted while the test was running.

Also the test for `File.rename` behavior would fail on Windows, because
a file cannot be renamed while any open handles to the file remain.

This commit changes the tests to use the PuppetSpec::Files methods for
generating a unique, temporary, and non-existent filename, and ensures
that any file/dir created during the test will be deleted, even if the
test fails during the run.