(#22535) Cleanup of desc/hook for new settings
For both new backwards compatibility settings, add a note to the desc
that these are set automatically by the agent when needed, so no need
to set them explicitly.

Remove hook with warning message for report_serialization_format as the
agent now emits notice-level messages when it drops into backward compatibility
mode to work with an older master.