(#22535) Fix support for agent's status endpoint
Without this patch, if the agent's status endpoint were accessed,
it would report the master's version.  This would happen because
the status indirection terminus was set to :rest for the duration of
an agent's lifetime.

This patch removes the puppet setting for status_terminus, and restores
the default terminus of local.  It also moves the check for master version
out of Puppet::Configurer (where it was shared with, say, apply) and into
Puppet::Agent.  And in the agent, the check *temporarily* changes the
endpoint to rest, queries the master's status endpoint, then restores the
original setting for the endpoint.

This patch also updates the corresponding spec tests, moving tests from
configurer_spec.rb to agent_spec.rb, and adding a stub to local_spec.rb.