(#22535) Make a full status object the default in agent spec tests.
Prior to this patch, the default stubs for the find indirection returned
a bare Puppet::Status object, i.e. without setting any version field in the hash,
and the one test that needed a version field overrode that stub with one
providing a status object with a version field.

This is the opposite of what happens in the code, where by default one would
expect Puppet::Status to include the current puppet version, and it's an unusual condition
(older master) for it to omit the version field.  So this patch reverses things
so that the default stub of Puppet::Status includes a version field, and the
two tests with specific expectations about the version field override the stub.