(#22529) Update acceptance test for regex fix and solaris
Confine solaris provider which does not implement purgeable, and place
the confine around the whole agents apply block so that it actually runs
on a confined subset of hosts.

Also fix the too restrictive regex which was missing host names with
hyphens and was insensitive to escape sequences (escape sequences are
very sensitive by nature and easily offended).