(#16384) Make puppet parser validate exit more cleanly
Users running puppet parser validate on a manifest with errors would see the appropriate errors in their manfiest, but would also be told to run "puppet help parser validate" for proper usage of the command indicating they ran it wrong when in fact they had not.  This patch causes puppet parser validate to only display the validation errors and not display the help message during a valid run of this face.