(#22324) Reset confine method to confine_block
confine_block is defined in my fork of Beaker.  confine will work here,
but only be accident, because a block passed to the confine method is
expected to further refine acceptable hosts.  confine_block just uses
the criteria to refine hosts and then executes the block on those hosts.

This commit is only in the integration/acceptance branch until we get
that merged in.