(#22652) Restore rack parsing of POST body parameters
The first pass at the fix for #22652 fixed ignore handling by fixing
support for multi-valued parameters in a GET query_string (i.e. by returning
them as an array); however, it inadvertently broke parsing of parameters
in a POST body, e.g. as found in the catalog endpoint POST.

This patch addresses parameters in either case.  As a footnote however,
note that multi-valued parameters in a POST body (which we don't use) would not
return them as an array.

This patch also adds two specs tests, one to catch the issue in the first pass fix
(POST parameters not parsed), and one to flag the state of POST multi-valued param
parsing (does not return array).