(#22646) Add testing of hiera.yaml version 3
This adds testing of hiera-2 hiera.yaml using version 3.
It also adds the issues introduced by the version 3 checker, and
corrects several issues in the checker implementation.

Fixture data reorganized (remove one level in the directory structure
and making it easier to remove support for version 2 data in the future.