(#22557) Raise an error providing a package resource with multiple names
Redmine #22557 brought up a regression in package provider behavior
(specifically dpkg).  Previously it was possible, under some
circumstances to supply a package resource with an array for the name
attribute, and have puppet manage the set of packages.  However, this
was an accidental feature which could fail if some packages were in
different states because the providers' query methods ultimately operate
on the package property hash info of just one of the packages.

This commit makes explicit that a package resource should not have its
name attribute set to an array of more than one value, and will raise a
ResourceError if this is attempted.