(#22180) Provide doc --mode rdoc parsing under Ruby 1.9
Puppet's RDoc parser was only compatible with RDoc 1, which is part of
Ruby 1.8.  RDoc's internals and API change between version 1 and 2+.

This commit refactors the rdoc/parser into an
rdoc/parser/puppet_parser_core.rb module which is included by two
different rdoc1 and rdoc2 parser classes.

It allows generation in either Ruby 1.8 or Ruby 1.9, but the HTML
renderer under Ruby 1.9 is stll the default, hence the format is not
optimal, and some elements, notable plugins and nodes, are not present.