(22740) create_resouces improved input validation
The create_resources() function in puppet expects a hash as its second
argument.

Currently, the only validation of input we perform is for number of
arguments. If the second parameter is not a hash, is nil, etc, Puppet
explodes. For example, you might get a message like `undefined method
`each' for "":String at /etc/puppetlabs/puppet/manifests/site.pp:45 on
node pe-301-master.puppetdebug.vlan`. Not very useful.

This patch adds a check that the second parameter is a hash, and raises an
error if it is not.