(#22593) Combine LBRACE handling into single rule
The LBRACE handling used to be split into two parts which made seeing
what an '}' character meant harder to follow and caused some duplicate
code, which only executed dependent on the order in which the lexer
tried to match the tokens. This combines the two cases together and
removes unneeded conditions.