(#19955) Move error handling closer to hiera
The error handling shouldn't be done inside Puppet::Resource, since most
of that handling was very specific to YAML, which is not the concern of
that code. This moves the error handling closer to hiera and removes
other errors that are not required to handle hiera errors. This should
deal with most errors for JSON, YAML, and general hiera problems.