(maint) Avoid accidental outdenting when appending stuff to doc fragments
When appending:

* Value collection summaries to parameter docs
* Dochook blurbs to provider docs

...we would sometimes cause an "outdent" if the original fragment ended with a
newline. Like so:

First line
    hanging indent
    hanging indent
  appended blurb

When we ran scrub() on that fragment later, we would get the wrong amount of
space removed. This silently sorta succeeded when scrub() was being lazy about
finding least common indent, but making scrub() work as expected made any sloppy
appends break the formatting.

This commit adds some scrub calls and makes our appends more predictable. This
will result in a certain amount of double-scrubbing, but that's okay.