(#22324) Update cve-2013-4761_* tests for current Beaker
These two security acceptance tests had been held back waiting for the
security patch to be released and available for several weeks.  We've
now added them back in, but their syntax predates the switch to Beaker.
this patch updates them to work with the current with_puppet_running_on
syntax.