(#22324) Fix modules/install/already_installed tests for pe
These two tests check the behavior of the module install tool when the
module we request to install is already installed.  The tests were
simulating an installed module by manually creating some files, but this
misses the step of chowning files based on the owner/group settings of
the moduledir, which is something the puppet module tool does, and
something module_utils is explicitly testing.

Rather than trying to simulate the owner/group setting that puppet
module tool would perform, I start off by installing the module with the
module tool, and then the test continues as before by attempting to
reinstall.