(maint) Use default rspec formatter for Travis CI
Without this commit, Travis CI can't display the full output of rspec
because Travis CI won't display the output of extremely long test runs.
This is problematic because a build can fail in Travis CI but the actual
failure in a broken build can be very hard to find and link to.

This commit changes the spec command to use the default formatter so
that the rspec output is less verbose and we can easily find and link to
failures in the Travis CI build.