(#22324) Confine puppet manages own configuration test not to run on pe
There are a few changes to better determine user and group from puppet
configuration, but at the end of the day, a pe installation also manages
the pe-puppet user/group for critical resources like the document root
used by apache/passenger.  Puppet knows nothing about this, so
destroying and recreating the pe-puppet user/group is not tenable for a
pe installation and I'm confining this test not to run on pe hosts.