(#22324) Puppet cert generate tests do not clear master ssl dir
A PE installation's ssl settings may not be easily cleared and recreated
(for the master node) as there are other components (passenger for
instance) which rely on the current settings.  This commit changes the
puppet_cert_generate_and_autosign.rb so that it does not delete and
recreate the master ssldir.  Agents are still cleared and reset.