(maint) Check jenkins queueItem as well for a pending job
The ci:check_for_downstream task would fail to find a job if it was
queued behind a currently running job as this are not listed in the
'builds' section.  The first queued item is listed under queueItem, and
this change checks there.  I have not found a way to query jobs queued
beyond the first.