(#22878) Scope call to CloseHandle
Previously, the win32-process gem used to extend Windows::Handle from the
windows-pr gem, so that the Process.CloseHandle method was available.
However, newer versions of the win32-process gem that are based on ffi, do
not extend Windows::Handle, so the Process.CloseHandle method is not defined.

This commit modifies puppet to call a version of CloseHandle which we
control, and therefore is guaranteed to exist.