(#22729) Remove the reject function.
The iterative 'reject' function shadowed the stdlib function with the
same name. At the same time there was a decision to rename the 'select'
function to 'filter' (to use a name in the functional programming
languages family map/reduce/filter). The reject function was then
rejected [sic].

This commit removes the reject function and the corresponding tests. The
function was not used anywhere else in puppet logic or testing.

User documentation needs to be updated.